requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
langchain==0.0.267
langchain-core==0.1.15
PyPDF2==3.0.1
python-docx==0.8.11
markdown==3.5
beautifulsoup4==4.12.2
pydantic==2.4.2
pytest==7.4.3
typing-extensions>=4.5.0
starlette>=0.27.0
pytest==7.4.3
pytest-mock==3.12.0
httpx==0.27.0   